endpoint.py is now renamed cli.py endpoint_manager.py is renamed endpoint.py EndpointManager is renamed Endpoint
